requirements.txt
ollama
pymongo
python-dotenv
neonize
schedule
